requirements.txt

openai~=1.9.0
fastapi~=0.109.0
starlette~=0.36.0
python-dotenv~=1.0.1
passlib~=1.7.4
pydantic~=2.5.3
pymongo~=4.6.1
python-jose[cryptography]
uvicorn~=0.26.0
click~=8.1.7
typing-extensions~=4.9.0
h11~=0.14.0
annotated-types~=0.6.0
pydantic-core~=2.14.6
httpx~=0.26.0
sniffio~=1.3.0
idna~=3.6.0
email-validator~=1.1.3
flask
psycopg2
bcrypt
python-multipart
fastapi_login
boto3
ultralytics
mlflow